Allow keywords in named import/export syntax
Progress toward #98

See the imports and exports grammar from the spec:
https://www.ecma-international.org/ecma-262/8.0/index.html#sec-imports
https://www.ecma-international.org/ecma-262/8.0/index.html#sec-exports

In particular, ImportSpecifier and ExportSpecifier use IdentifierName to
identify the import/export name rather than the more restrictive ImportedBinding
(or just Identifier). That means any keyword is allowed.

There are (at least) three contexts to think about in JS identifier parsing:
* In most contexts, identifiers like variable names can be any name except a
  reserved word.
* Some keywords, like "as", are normally not keywords but have special
  keyword-like meaning in specific situations. Acorn and Babel call these
  "contextual keywords", and `tokenIsAlsoIdentifier` and `preferIdentifier` in
  AssemblyScript appear to handle that case.
* In some contexts, like object keys and named imports/exports, any name is
  allowed, even keywords.

This change adds that third case in order to handle imports/exports. I added a
`forceIdentifier` flag that tells the tokenizer that any identifier name should
be treated as a plain identifier, then used it in the import and export cases.