Add a more helpful error message when you haven't defined an allocator
This tripped me up, and it looks like others have run into it as well, so it
seems like a good idea to have a custom error message.

I also added a system for writing tests that assert that certain error codes are
triggered so that I could test this.