feat(examples/n-body) Fix and improve the Rust example.
This patch first updates the documentation to compile the Rust example.

Then this patch updates the Rust code, since `panic_implementation` is
a deprecated feature. Also `extern` is similar to `extern "C"`. Less
code is always better.

Finally, this patch compiles and optimises the Rust Wasm binary file
(with a simple `wasm-gc` run). The final binary size is 1.4Kb.