Enforce strict field initialization (#1349)
BREAKING CHANGE: Instance fields not definitely assigned within the constructor now yield an error, unless the field's name is annotated with a trailing `!` (checked at runtime), is of a nullable or is of a value type (initialized to zero). Fields of non-nullable reference types must be assigned before `this` is either implicitly or explicitly returned from the constructor, passed to another function or another instance method is called on `this`. Overall the behavior is similar to, but not exactly the same as `--strictPropertyInitialization` in TypeScript.