Add global types recognition for advanced developers
When dealing with adding global types for end users, as-pect and cli tools have always relied on adding something like this to an end user's project:

```ts
/// <reference path="../path/to/types.d.ts" />
```

I instead propose standardizing recognizing types that follow the `../node_modules/**/assembly/global.d.ts` pattern out of the box. This of course is not a perfect solution, and I am amenable to other solutions in the same vein as this one, however, this one seems like the simplest way to fix a long standing problem for cli developers using the compiler programatically.