Allow for generics/parameter types in inner types
This feature takes a somewhat hacky approach to adding these types to
the Flow. Each Flow now stores a Map of strings and TypeDefinitions, and
most of the resolver's methods propagate this map when resolving types
inside functions. The TypeDefinitions are created by the compiler upon
encountering a TYPEDECLARATION node, just like inner types without
parameters.

The existing code handling TypeDefinitions in the resolveNamedType()
method was abstracted into a resolveTypeDefinition() method. That way,
that code can be reused to resolve contextual TypeDefinitions in the
Flow. However, there is a bug: an inner type with a parameter
cannot reference another inner type or the outer function's type
parameter. Commenting out `ctxTypes.clear();` in the Resolver's
resolveTypeArguments() method seems to fix this issue.