Bump the lockfile
The latest packages that were added to the lockfile when resolving a
rebase conflict might not have survived the rest of the rebase, so
another bump is provided here instead.

This bump also caused TypeScript to emit an error related to an
undici-types package due to this bump, and that is also fixed here.