BREAKING CHANGE: Avoid collisions between static class/namespace members
This commit changes the delimiter between namespaces and their members
from "." to "::" in internal names. The new delimiter certainly feels
very un-JavaScript-y, but the meaning is still recognizable, and these
names aren't exposed to users in most cases anyway. Still, this is a
breaking change, as it affects the import names for `declare namespace`
members (as seen in tests/compiler/declare.js), and transform authors
will also need to update their code accordingly.

Fixes #2793.