Forbid duplicate static class and/or namespace members
This isn't fully TS compatible, but refactors targeting internal names,
scoping, merging, etc. are needed to become more compatible. For
instance, if namespace members had unique separators in internal names,
then a non-exported namespace member would override a static class
member, assuming the names are the same.

Note that this change doesn't prevent the compiler from attempting to
compile the duplicate global, and hence the previous commit is needed
for this to work fully.

Barely fixes #2793.