Mark the module as closed-world when possible to allow DAE optimizations
Due to WebAssembly/binaryen#7135, any function reference is treated as
potentially leaked to the outside world, unless the module is marked as
closed-world. This prevents DAE from taking place. To fix this, mark
the module as closed-world when tables are neither imported nor exported
and when exports do not return/contain funcrefs.